Build Get Department REST API

+++ src/main/java/net/javaguides/emsbackend/repositories/IDepartmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IDepartmentRepository extends JpaRepository<Department, Integer> {
+public interface IDepartmentRepository extends JpaRepository<Department, Long> {
 }
